+++ src/MicroElements.Metadata/Validation/Rules/NotNull.cs
@@ -17,7 +17,7 @@ namespace MicroElements.Validation.Rules
         /// </summary>
         /// <param name="property">Property to check.</param>
         public NotNull(IProperty<T> property)
-            : base(property, "Property {propertyName} should not be null")
+            : base(property, "{propertyName} should not be null")
         {
         }
 
@@ -49,6 +49,18 @@ namespace MicroElements.Validation.Rules
         /// Checks that property value is not null.
         /// </summary>
         /// <typeparam name="T">Property type.</typeparam>
+        /// <param name="property">Property to check.</param>
+        /// <returns>NotNull validation rule.</returns>
+        public static NotNull<T?> NotNull<T>(this IProperty<T?> property)
+            where T : struct
+        {
+            return new NotNull<T?>(property);
+        }
+
+        /// <summary>
+        /// Checks that property value is not null.
+        /// </summary>
+        /// <typeparam name="T">Property type.</typeparam>
         /// <typeparam name="TValidationRule">Combined validation rule type.</typeparam>
         /// <param name="linker">Rule linker.</param>
         /// <returns>Combined validation rule.</returns>
@@ -57,6 +69,20 @@ namespace MicroElements.Validation.Rules
             where T : class
         {
             return linker.CombineWith(new NotNull<T>(linker.FirstRule.Property));
+        }
+
+        /// <summary>
+        /// Checks that property value is not null.
+        /// </summary>
+        /// <typeparam name="T">Property type.</typeparam>
+        /// <typeparam name="TValidationRule">Combined validation rule type.</typeparam>
+        /// <param name="linker">Rule linker.</param>
+        /// <returns>Combined validation rule.</returns>
+        public static TValidationRule NotNull<T, TValidationRule>(this IValidationRuleLinker<T?, TValidationRule> linker)
+            where TValidationRule : IValidationRule<T?>
+            where T : struct
+        {
+            return linker.CombineWith(new NotNull<T?>(linker.FirstRule.Property));
         }
     }
 }